Added update board, player and test files
The new player file now contains new calculateGameTreeWithMinMax and other related functions that combine the properties of the minmax algorithm with the calculateGameTree function. The tree is now minimaxed as it is built, instead of built entirely and iterated over in another function.
This convolution was a step towards building calculateGameTreeWithMinMax2 which uses Alpha-Beta Pruning to streamline the algorithm. Instead of building a full tree and traversing it to minimax and pruning where appropriate, this approach stops the construction of prunable branches in the first place. 

The board file was modified to include a test board for the new minimax algorithm. 
The test file was used for debugging.

Overall I have observed some improvements in speed that seem to increase with the ply of the search, which is as expected and good news.
However, a bug has been discovered: the tree search algorithm assumes a move is equivalent to the end of a turn, which of course in checkers is often not the case. Handling multiple moves in a turn will will be my next task.