Added consructAlphaBetaPrunedTree via upload
In my previous update, I outlined the bug in the existing alpha-beta pruned minimax algorithm I had written: that it does not account for multiple moves in a single turn in the event of several pieces being taken. 

Fortunately, it seems a simple solution exists. If I shift the dependancy of max or min from wither the move is finished to which players turn it is (which can be extracted from the state of the board at  particular node) then the algorithm should function correctly. 
To take steps toward this, I decided to merge the two functions that alternately recursed into a single function with a conditional statement inside. The conditional statement currently still changes based on a transition, but the idea is that this conditional can then be modified.

While rewriting, I abstracted and condensed the syntax, and now the function more closely resembles the Pseudo code from the Wikipedia article https://en.wikipedia.org/wiki/Alpha%E2%80%93beta_pruning . 

This new file I am uploading is the general function and is tested on a minimal state class object defined at the top of the program. 

My next task will be integrating this format back into my 'player' file and implementing the dependency on who's turn it is.