Clean-up of labelling conventions.
It's been a little while, but much progress has been made. Since my last update, I have analysed the Board state class carefully and discovered not two, but three conventions being used for labelling squares! What a nightmare. I have solidified a new strategy towards labelling which involves a Reading convention (R-convention) for human input and output to the screen for a human ONLY, and a Programming convention (P-convention) which is for everything else.
I have detailed the comments and the docstrings in the code to reflect my more fundamental understanding and also to make the structure of the program make more sense.
There is still work to be done. The way that turns are changed via the move function is very clunky and unintuitive, involving the 'change_sides' fucntion which absolutely does not do what it is named. I have developed a strategy to change this so that it is more pythonic, and will work as intended.
Furthermore, the conversion from numpy-array to binary representiation seems theoretically correct, but is definitely still not functioning correctly. It will require some more sophisticated debugging, which I will investigate soon. I am looking forward to getting some answers.