Change of turn reworked. Learning update made into seperate function.
The changing of turns has been reworked so that it is more intuitive. Fortunately this allowed me to removed the change_sides function entirely from the object which, as discussed previously, was confusingly named.
Instead, I have made liberal use of the object property turn to correctly calculate which pieces are the current player's and which are the opponenets.
Furthermore, I have extracted the learning parameter update from the function the selects a move, and put it in a new function entirely. This allows it to be called at any time needed. The motivation was that, since the update was only happening when it was the TD Learner's turn, it was not able to update it's weights when it lost as it did not get the oppertunity to have another turn.
My initial fix was to update the learner at the end of every turn, but this caused more problems. The present update contains these problems.
I realised that I had not understood the theorey correctly, and the states between which the updates are to be made are not any consecutive game state, but specifically consecutive game states in which it is also the agent's turn. Next, I will be adjusting the updates to incoporate this correctly. I will also be adding a special case for the end of the game to allow a losing agent to learn correctly.